core: arm32: disable interrupt in thread_excp_vect_workaround
thread_excp_vect_workaround isn't reentrant because it use the tpidr
as a temporary register to save value of r0. That means if a fiq
happend when optee is processing a syscall, the syscall argument r0
will be changed to unexpected value.

Move `write_tpidrprw r0` out of `vector_prologue_spectre` and add
`cpsid aif` before it to fix this.

Signed-off-by: Mark-PK Tsai <mark-pk.tsai@mediatek.com>